requirements.txt
streamlit>=1.30.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
faiss-cpu>=1.7.0
